VZ-1306: Move testutils to separate package
